pyproject.toml
[project]
name = "syp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django-rest-knox>=5.0.2",
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "django-debug-toolbar>=4.4.6",
]
